[Decomp][Player] Cohesion split: extract health/power regeneration to PlayerRegen.cpp
